Added PDF output
Josh Cummings was able to puzzle out the classpath collision that had hung up producing PDF, so now we have PDF. Epub output causes a test to fail when the install task is run but not when the test is run by itself. Because of the failing test, I backed out epub output, at least for now.